app/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
Faker==27.4.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
instructor==1.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.3.7
langfuse==2.27.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
openai==1.42.0
orjson==3.10.5
packaging==23.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv
python-frontmatter==1.1.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.1
pyzmq==26.0.3
redis==5.0.3
requests==2.32.3
rich==13.7.1
sentry-sdk==2.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
tenacity==8.5.0
tokenizers>=0.21,<0.22
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
timescale-vector==0.0.7
pandas==2.2.3
graphviz==0.20.3 
pandoc
torch==2.6.0
transformers==4.50.3
pymilvus==2.5.5
PyMuPDF==1.25.4
python-docx==1.1.2
elasticsearch==8.11.0
minio==7.1.14
anthropic==0.49.0
instructor==1.7.9
pydantic-settings==2.8.1
sentencepiece==0.2.0
pymilvus==2.5.6
torch==2.6.0
transformers==4.51.3
